+++ api/controls/serializers/element.py
@@ -32,14 +32,11 @@ class DetailedElementSerializer(SimpleElementSerializer):
         fields = SimpleElementSerializer.Meta.fields + ['roles', 'import_record', 'tags']
 
 class ElementPermissionSerializer(SimpleElementSerializer):
-    # users_with_permissions = get_users_with_perms(element, attach_perms=True)
     users_with_permissions = serializers.SerializerMethodField('get_list_of_users')
-    # users_with_permissions = 'perm1'
-
+    
     def get_list_of_users(self, element):
-        users_dict = {} #user_id: permissions
-        # running get_users_with_perms(element, attach_perms=True) gives me a queryset, i cant return a query set, so i need to 
-        #transform the query set to a dict with the user id and the perms to that id
+        users_dict = {} # user_id: permissions
+        # Transform get_users_with_perms(element, attach_perms=True) queryset to a dict user id and  perms
         userPermsQS = get_users_with_perms(element, attach_perms=True)
         for user in userPermsQS:
             listOfPerms = list(get_user_perms(user, element))